feat: rename "db.py" to "storage.py"
