+++ deploy/dynamo/api-server/db/start_db.py
@@ -19,7 +19,7 @@ import os
 
 import uvicorn
 from db.api import app
-from db.db import create_db_and_tables_async
+from db.storage import create_db_and_tables_async
 
 logger = logging.getLogger(__name__)
